drivers/nutdrv_qx.c: range-check and cast "u_langid_fix" in upsdrv_initups()

+++ drivers/nutdrv_qx.c
@@ -2497,10 +2497,12 @@ void	upsdrv_initups(void)
 		if (getval("langid_fix")) {
 			/* Skip "0x" prefix and set back to hexadecimal */
 			unsigned int u_langid_fix;
-			if ( (sscanf(getval("langid_fix") + 2, "%x", &u_langid_fix) != 1) || (u_langid_fix > INT_MAX)) {
+			if ( (sscanf(getval("langid_fix") + 2, "%x", &u_langid_fix) != 1)
+			||   (u_langid_fix > INT_MAX)
+			) {
 				upslogx(LOG_NOTICE, "Error enabling language ID workaround");
 			} else {
-				langid_fix = u_langid_fix;
+				langid_fix = (int)u_langid_fix;
 				upsdebugx(2, "Language ID workaround enabled (using '0x%x')", langid_fix);
 			}
 		}
